flask_impl/requirements.txt
feedparser==5.2.1
Flask==1.0.2
newsapi-python==0.2.3

